chore(lint): enable unconvert (#213)

+++ util/inodeutil/inode_other.go
@@ -12,7 +12,7 @@ func GetInodeBestEffort(path string) uint64 {
 	var stat syscall.Stat_t
 	inode := uint64(0)
 	if err := syscall.Stat(path, &stat); err == nil {
-		inode = uint64(stat.Ino)
+		inode = stat.Ino
 	}
 	return inode
 }
